made changes based on the review

+++ src/main/resources/init_db.sql
@@ -1,5 +1,5 @@
 CREATE TABLE `books` (
-  `id` int NOT NULL AUTO_INCREMENT,
+  `id` BIGINT(255) NOT NULL AUTO_INCREMENT,
   `title` varchar(255) DEFAULT NULL,
   `price` decimal(10,2) DEFAULT NULL,
   PRIMARY KEY (`id`)
